JAV-123 removed duplicate tests, since we are tapping on invocation
