+++ coverage-reports/pom.xml
@@ -240,12 +240,7 @@
     </dependency>
     <dependency>
       <groupId>io.servicecomb.tests</groupId>
-      <artifactId>tracing-servlet-tests</artifactId>
-      <version>0.1.0-m3-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>io.servicecomb.tests</groupId>
-      <artifactId>tracing-vertx-tests</artifactId>
+      <artifactId>tracing-tests</artifactId>
       <version>0.1.0-m3-SNAPSHOT</version>
     </dependency>
   </dependencies>
