--- integration-tests/test-common/src/test/java/io/servicecomb/tests/tracing/Log4jConfig.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2017 Huawei Technologies Co., Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.servicecomb.tests.tracing;
-
-import org.apache.log4j.Appender;
-import org.apache.log4j.Level;
-import org.apache.log4j.Logger;
-
-public class Log4jConfig {
-
-  public static void addAppender(Appender appender) {
-    Logger rootLogger = Logger.getRootLogger();
-    rootLogger.removeAllAppenders();
-    rootLogger.setLevel(Level.INFO);
-    rootLogger.addAppender(appender);
-  }
-}
