+++ integration-tests/tracing-tests/src/test/java/io/servicecomb/tests/tracing/TracingTestMain.java
