+++ integration-tests/tracing-tests/src/test/resources/microservice.yaml
