+++ integration-tests/tracing-tests/src/test/resources/registry.yaml
