--- integration-tests/tracing-tests/tracing-servlet-tests/src/test/webapp/WEB-INF/web.xml
@@ -1,48 +1,0 @@
-<!--
-  ~ Copyright 2017 Huawei Technologies Co., Ltd
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~    http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<web-app xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
-          http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
-  version="3.0">
-
-  <context-param>
-    <param-name>contextConfigLocation</param-name>
-    <param-value>
-      classpath*:META-INF/spring/*.bean.xml
-      classpath*:app-config.xml
-    </param-value>
-  </context-param>
-  <context-param>
-    <param-name>servicecomb.logging.merged</param-name>
-    <param-value>false</param-value>
-  </context-param>
-
-
-  <listener>
-    <listener-class>io.servicecomb.transport.rest.servlet.RestServletContextListener</listener-class>
-  </listener>
-  <servlet>
-    <servlet-name>RestServlet</servlet-name>
-    <servlet-class>io.servicecomb.transport.rest.servlet.RestServlet</servlet-class>
-    <load-on-startup>1</load-on-startup>
-    <async-supported>true</async-supported>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>RestServlet</servlet-name>
-    <url-pattern>/</url-pattern>
-  </servlet-mapping>
-</web-app>
